// app/model/Food.tsx
export interface Food {
    id: string;            // ID của đồ ăn
    name: string;          // Tên của đồ ăn
    price: number;         // Giá của đồ ăn
    quantity: number;      // Số lượng đồ ăn
    sold: number;          // Số lượng đã bán
    imageUrl: string;      // Link ảnh của đồ ăn
}